Fix: attach sidebar and mainview between sm and upper size
