+++ src/app/(main)/home/search/[keyword]/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <div className="flex flex-col h-full bg-black-121212 text-white">
+      <h1>bye!</h1>
+    </div>
+  );
+}
