+++ src/app/_components/(main)/home/serverComponents.tsx
@@ -29,7 +29,7 @@ export function NowPlaying() {
 // TODO: Scrollbar CSS 처리
 export function MainView({ children }: { children: React.ReactNode }) {
   return (
-    <div className="h-full grow rounded-lg overflow-y-auto md:pl-[19.5rem]">
+    <div className="h-full grow rounded-lg overflow-y-auto sm:pl-[19.5rem]">
       <div className="bg-black-121212 text-white h-full flex flex-col">
         <div className="flex-grow p-4 overflow-y-auto scrollbar-hide">
           {children}
@@ -41,7 +41,7 @@ export function MainView({ children }: { children: React.ReactNode }) {
 
 export function SideBar() {
   return (
-    <div className="hidden h-full sm:block bg-black-121212 fixed left-0 right-auto w-[19rem] p-5 z-20 rounded-lg">
+    <div className="hidden h-full sm:block bg-black-121212 fixed left-0 w-[19rem] p-5 z-20 rounded-lg">
       <aside aria-label="Left Sidebar Navigation" className="relative">
         <div className="flex flex-row gap-3 items-center p-5 font-lg font-bold text-white">
           <LibraryLogo />
@@ -49,9 +49,9 @@ export function SideBar() {
           <PlusLogo />
         </div>
         {/* {
-                isLogin ? <Library />
-                : <DummyLibrary />
-                } */}
+          isLogin ? <Library />
+          : <DummyLibrary />
+        } */}
       </aside>
     </div>
   );
